pyproject.toml
[tool.poetry]
name = "searchable-pdf"
version = "0.1.0"
description = ""
authors = ["nogara <nogara@motrix.global>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
PyPDF2 = "^3.0.1"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
